refactor(Fieldets): seperate field components
